Refactor classes
made HypermediaRequest class as common CRUD base class for Protocol
specific request generators
